fix: Add code missing from App Tools patch
This commit fixes a missing code issue in the `patchNow` function of the
App Tools modal. The issue was that the code was referencing a variable
called `package` which wasn't actually being passed in the argument
object.

This commit updates the code to use the correct variable name `pkg`,
which is now being passed in the argument object. This ensures that the
App Tools modal functions correctly and can access the correct
information about the application being managed.