build: Add markdownlint
These changes introduce Markdown linting to enforce consistent
formatting and style in the project's Markdown files.

The `markdownlint` and `markdownlint-cli` packages were added to the
project's dependencies, and a configuration file `.markdownlint.jsonc`
was created to specify the desired Markdown linting rules.  The
`markdownlint` command was added to the `package.json` scripts to enable
linting as part of the build process.

The addition of Markdown linting helps to ensure that the project's
documentation is consistently formatted, improving readability and
maintainability.